Refactor resolveAttack test coverage
